rename shouldAddLimit func
